No longer automatically protect they key column
Rails 4 no longer supported whitelisted columns for protection. That
functionality has been extracted into a Gem that we don't want as a
dependency for this gem, so we're taking out the functionality and
leaving it to the user to protect their columns.

This is also good because it gives the user full control over what
is and isn't protected by default.